Log levels / XML auto-export / Forced export and import.
This commit lowers the log level for ExportRoutine() method where the export
message wasn't being show as expected. This commit also adds two new
features as described below:

1. The possibility to auto-export the XML when making project changes.

This allows the developer to skip the need the keep the XML up-to-date
with the plain source code.

2. The possibility to force the project to be exported/imported.

Now it's possible to force the project export/import using the
Source control menu. Beware though, since there isn't any way to confirm
the action.