Lots of changes.
1. Added the possibility to ignore paths whenever a file called .portignore
composed is present in the repository. All files or directories inside those
paths are also ignored.

2. Removed the auto compilation function to improve stability by reducing the locks chances.
Now the user is responsible for compile their projects.

3. Simplified the message reporting how many items were affected when
importing a project.

4. When installing port (using the Installer), it will map ^Port.SourceControl.Projects and ^Port.SourceControl.Settings to the namespace
that called the installer. This is done to assure that the global being used won't be mapped to another namespace.

4. Fixed an issue where ExportToXML was being called multiple times when importing a
XML using the native Studio tool.