More a bunch of fixes and features (detail this commit).
Features:

- The introduction of Port.REST.API

This commit introduces a class called Port.REST.API that actually uses Atelier behinds the scenes.
It's not usable for now, but it can provide some idea about how it could work.
Think about this class as being some kind of an adapter, that receives projects and translates into scoped items.
Also, it's decided to not make Port aware of Eclipse Atelier, but instead the Atelier REST API, this allows multiplee editor implementations, including VSCode and Atom.

- Support for custom source file extensions

Due to a request it's now possible to change the source code extension by using the class method SetSourceExtension from the class Port.SourceControl.Config.
It also accepts an empty string, which keeps each routine's original extension.

Note that changing the extension doesn't affect the exported files if they're exported atomically.
It's recommended to force export the project if the user wishes to keep the repository clear.

Bugs fixed:

- Importer is aware of the ignored paths

This commit improves the usage of the ignore feature (by using .portignore file),
making the importer also aware of the ignored paths.
This creates a policy where the user must keep the possible builds (bundles) in sync with the projects.

- Refactored extraneous item search algorithm for better coverage

This fixes an empty directory regression bug.